Frontend-monorepo-45: Removed unnecessary refactor added as part of layout test
