+++ apps/token/src/routes/staking/index.tsx
@@ -1,12 +1,32 @@
 import React from 'react';
-import { Outlet } from 'react-router-dom';
+import { useTranslation } from 'react-i18next';
+import { Outlet, useMatch } from 'react-router-dom';
+
+import { Heading } from '../../components/heading';
 import { useDocumentTitle } from '../../hooks/use-document-title';
 import type { RouteChildProps } from '..';
 
 const StakingRouter = ({ name }: RouteChildProps) => {
   useDocumentTitle(name);
+  const { t } = useTranslation();
+  const associate = useMatch('/staking/associate');
+  const disassociate = useMatch('/staking/disassociate');
 
-  return <Outlet />;
+  const title = React.useMemo(() => {
+    if (associate) {
+      return t('pageTitleAssociate');
+    } else if (disassociate) {
+      return t('pageTitleDisassociate');
+    }
+    return t('pageTitleStaking');
+  }, [associate, disassociate, t]);
+
+  return (
+    <>
+      <Heading title={title} />
+      <Outlet />
+    </>
+  );
 };
 
 export default StakingRouter;
